requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
tweepy==4.14.0
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
sqlalchemy==2.0.23
pydantic==2.5.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cors==1.0.1